refactor: extract duplicated functions from map components
