+++ src/screens/prediction/components/prediction-map/component.js
@@ -13,19 +13,22 @@
 import React, { useState, useEffect } from 'react';
 import ReactTooltip from 'react-tooltip';
 import mapboxgl from 'mapbox-gl';
-import printPdf from 'mapbox-print-pdf';
 
 import {
   DATA_MODES,
   stateAbbrevToZoomLevel,
-  MAP_SOURCES,
   SOURCE_LAYERS,
   MAP_SOURCE_NAME,
   VECTOR_LAYER,
   STATE_VECTOR_LAYER,
+  MAP_TYPES,
 } from '../../../../constants';
 
-import { getMapboxRDNameFormat, getFillColor } from '../../../../utils';
+import {
+  getMapboxRDNameFormat, getFillColor, createMapClickCallback, generateMap,
+  downloadMap,
+  createHoverCallback,
+} from '../../../../utils';
 import { api } from '../../../../services';
 
 import {
@@ -84,36 +87,13 @@ const PredictionMap = (props) => {
     }
   };
 
-  // twice-curried function for generating hover callback
-  const createMapHoverCallback = (predictions, rangerDistricts, mode, state, availStates) => (e) => {
-    if (!map || !e || !map.isStyleLoaded()) return;
-
-    const counties = map.getLayer(VECTOR_LAYER)
-      ? map.queryRenderedFeatures(e.point, { layers: [VECTOR_LAYER] })
-      : [];
-
-    if (counties.length > 0 && counties[0]?.properties?.forest) {
-      const { x, y } = e.point || {};
-
-      const {
-        STATE: hoverState,
-        COUNTYNAME: hoverCounty,
-        forest: rawForest,
-      } = counties[0].properties;
-
-      // handles case where tileset has two spaces instead of one (this is a one-off), or is missing the word RD altogether (also one-off)
-      const hoverRD = rawForest.replaceAll('  ', ' ');
-
-      const location = mode === DATA_MODES.COUNTY
-        ? hoverCounty
-        : rangerDistricts.filter((rd) => !!rd)
-          .find((rd) => getMapboxRDNameFormat(rd)?.includes(hoverRD));
-
+  const createMapHoverCallback = (predictions, rangerDistricts, mode, state, availStates) => {
+    const callback = (hoverState, location, x, y) => {
       const pred = predictions.find((p) => {
-        // either ranger district mode or have a matching state
+      // either ranger district mode or have a matching state
         return (mode === DATA_MODES.RANGER_DISTRICT || (p.state === hoverState && p.state === state) || (!state && availStates.includes(hoverState)))
-          // and sublocation matches
-          && ((p.county === location && mode === DATA_MODES.COUNTY && p.state === hoverState) || (p.rangerDistrict === location && mode === DATA_MODES.RANGER_DISTRICT));
+              // and sublocation matches
+              && ((p.county === location && mode === DATA_MODES.COUNTY && p.state === hoverState) || (p.rangerDistrict === location && mode === DATA_MODES.RANGER_DISTRICT));
       });
 
       if (pred && x && y) {
@@ -135,97 +115,9 @@ const PredictionMap = (props) => {
       } else {
         setPredictionHover(null);
       }
-    }
-  };
+    };
 
-  // twice-curried function for generating click callback
-  const createMapClickCallback = (states, sublocations, currentState, predictions, mode) => (e) => {
-    if (!e?.features[0]?.properties) return;
-
-    const {
-      COUNTYNAME: county,
-      forest: clickRD,
-      STATE: mapboxState,
-    } = e.features[0].properties;
-
-    const rangerDistrictToSet = sublocations.filter((rd) => !!rd)
-      .find((district) => district.includes(clickRD));
-
-    const state = !mapboxState && mode === DATA_MODES.RANGER_DISTRICT
-      ? predictions.find((p) => p.rangerDistrict === rangerDistrictToSet)?.state
-      : mapboxState;
-
-    // ensure clicked on valid state
-    if (!states.includes(state) || !currentState) return;
-
-    // select county or RD depending on mode
-    if (dataMode === DATA_MODES.COUNTY && sublocations.includes(county)) {
-      if (props.county.length > 0) { // remove selection if user clicks selected county
-        setCounty([]);
-      } else {
-        setCounty([county]);
-      }
-    } else if (sublocations.includes(rangerDistrictToSet)) {
-      if (props.rangerDistrict.length > 0) { // remove selection if user clicks selected ranger district
-        setRangerDistrict([]);
-      } else {
-        setRangerDistrict([rangerDistrictToSet]);
-      }
-    }
-  };
-
-  const generateMap = (forceRegenerate) => {
-    if (map && !forceRegenerate) return;
-
-    const createdMap = new mapboxgl.Map({
-      container: 'map', // container id
-      style: 'mapbox://styles/pine-beetle-prediction/ckgrzijos0q5119paazko291z',
-      center: [-84.3880, 33.7490], // starting position
-      zoom: 4.8, // starting zoom
-      options: {
-        trackResize: true,
-      },
-    });
-
-    createdMap.addControl(new mapboxgl.NavigationControl());
-
-    const legendTagsToSet = thresholds.map((threshold, index) => {
-      const color = colors[index];
-
-      return (
-        <div key={color}>
-          <span className="legend-key" style={{ backgroundColor: color }} />
-          <span>{threshold}</span>
-        </div>
-      );
-    });
-
-    // add legend tags
-    setLegendTags(legendTagsToSet);
-
-    // add map source on load
-    if (!createdMap._listeners.load) {
-      createdMap.on('load', () => {
-        if (!createdMap.getSource(MAP_SOURCE_NAME)) {
-          createdMap.addSource(MAP_SOURCE_NAME, dataMode === DATA_MODES.COUNTY ? MAP_SOURCES.COUNTY : MAP_SOURCES.RANGER_DISTRICT);
-        }
-      });
-    }
-
-    // select county/RD when user clicks on it
-    if (!createdMap._listeners.click) {
-      const callback = createMapClickCallback(availableStates, availableSublocations, selectedState, data, dataMode);
-      setMapClickCallback(() => callback);
-      createdMap.on('click', VECTOR_LAYER, callback);
-    }
-
-    if (createdMap._listeners.mousemove === undefined) {
-      const callback = createMapHoverCallback(data, allRangerDistricts, dataMode, selectedState, availableStates);
-      setMapHoverCallback(() => callback);
-      createdMap.on('mousemove', callback);
-    }
-
-    setMap(createdMap);
+    return createHoverCallback(map, rangerDistricts, dataMode, callback);
   };
 
   const colorPredictions = (predictions) => {
@@ -289,105 +181,14 @@ const PredictionMap = (props) => {
     }
   };
 
-  // adopted from old site
-  // Creates and returns HTML with the title for the header
-  // of the downloaded maps. This object is used by the mapbox-print-pdf library.
-  const buildHeader = () => {
-    return (
-      `<div id="map-header" style="text-align: center;">
-          <h2 style="letter-spacing: 1px;margin-top: 200px;margin-bottom: 50px;">Probability of (Any) SPB Spots</h2>
-        </div>`
-    );
-  };
-
-  // adopted from old site
-  // Creates and returns HTML with information for the footer
-  // of the downloaded maps. This includes a legend for the color scale,
-  // notes explaining the legend and sources, and information about the
-  // data collection process. This object is used by the mapbox-print-pdf library.
-  const buildFooter = () => {
-    const title = `Southern Pine Beetle Outbreak Prediction Maps: ${selectedState} ${year}`;
-
-    // creates the color boxes and text fields for the legend in the footer
-    const legendString = thresholds.reduce((acc, curr, index) => {
-      const layer = curr;
-      const color = colors[index];
-      const spanString = `
-          <div class="footer-legend-key" style="font-family: 'Open Sans', arial, serif;background: ${color};display:
-          inline-block;border-radius: 20%;width: 20px;height: 20px;margin-right: 5px;margin-left: 5px;"></div><span>${layer}</span>`;
-      return acc.concat(spanString);
-    }, '');
-
-    return (
-      `
-          <div id="map-footer" style="text-align: center;letter-spacing: 1px;margin-top: 20px;margin-bottom: 0;">
-              <div id="footer-legend" style="font-family: 'Open Sans', arial, serif;width: 90%;margin: auto;margin-bottom: 10px;">
-                  ${legendString}
-              </div>
-              <p class="footnote" style="font-family: 'Open Sans', arial, serif;color: #898989;line-height:
-              14px;width: 53%;margin: auto;margin-bottom: 16px;font-size: 14px;">Note: Color ramp ascends with a constant factor of
-              increase in the probability of outcome.</p>
-              <div id="spacer" style="height: 50px;"></div>
-              <h2 style="font-family: 'Open Sans', arial, serif;margin-bottom: 16px;margin-top: 16px;">${title}</h2>
-              <p style="font-family: 'Open Sans', arial, serif;font-size: 14px;margin-bottom: 16px;">Predictions are based on a zero-inflated Poisson model fit to historical data
-              from 1988 – 2009 (Aoki 2017). The most important drivers of model predictions are
-               SPB trap captures in the current spring and SPB spots the previous year.
-              </p>
-              <p style="font-family: 'Open Sans', arial, serif;font-size: 14px;margin-bottom: 16px;">
-              The SPB prediction project is supported by USDA Forest Service: Science and Technology
-              Development Program (STDP)
-              </p>
-              <p style="font-family: 'Open Sans', arial, serif;font-size: 14px;margin-bottom: 16px;">Contact: Matthew P. Ayres - matthew.p.ayres@dartmouth.edu; Carissa F. Aoki - caoki@bates.edu
-              </p>
-              <p class="footnote" style="font-family: 'Open Sans', arial, serif;color: #898989;line-height: 14px;width: 53%;
-              margin: auto;margin-bottom: 16px;font-size: 14px;">Sources: Esri, HERE, Garmin, Intermap, increment P Corp., GEBCO, USGS,FAO, NPS, NRCAN,
-              GeoBase, IGN, Kadaster NL, Ordnance Survey, Esri Japan, METI, Esri China (Hong Kong), swisstopo, © OpenStreetMap
-              contributors, and the GIS User Community</p>
-          </div>
-          `
-    );
-  };
-
-  const downloadMap = () => {
-    if (!map || !year || isDownloadingMap) return;
-
-    setIsDownloadingMap(true);
-
-    const mapName = `${selectedState || 'All States'}-${year}.pdf`;
-
-    printPdf.build()
-      .header({
-        html: buildHeader(),
-        baseline: { format: 'a3', orientation: 'p' },
-      })
-      .footer({
-        html: buildFooter(),
-        baseline: { format: 'a3', orientation: 'p' },
-      })
-      .margins({
-        top: 8,
-        right: 8,
-        left: 8,
-        bottom: 8,
-      }, 'pt')
-      .format('a3')
-      .portrait()
-      .print(map, mapboxgl)
-      .then((pdf) => {
-        pdf.save(mapName);
-        setIsDownloadingMap(false);
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
-
   useEffect(() => {
     mapboxgl.accessToken = process.env.MAPBOX_ACCESS_TOKEN;
+    const clickCallback = createMapClickCallback(availableStates, availableSublocations, selectedState, data, dataMode, props.county, setCounty, props.rangerDistrict, setRangerDistrict);
+    const hoverCallback = createMapHoverCallback(data, allRangerDistricts, dataMode, selectedState, availableStates);
 
     setTimeout(() => {
       setMap(undefined);
-      generateMap(true);
+      generateMap(true, map, thresholds, colors, setLegendTags, dataMode, clickCallback, setMapClickCallback, hoverCallback, setMapHoverCallback, setMap);
     }, 100);
   }, [dataMode]);
 
@@ -435,7 +236,7 @@ const PredictionMap = (props) => {
       if (mapClickCallback) map.off('click', VECTOR_LAYER, mapClickCallback);
 
       // generate new callback
-      const callback = createMapClickCallback(availableStates, availableSublocations, selectedState, data, dataMode);
+      const callback = createMapClickCallback(availableStates, availableSublocations, selectedState, data, dataMode, props.county, setCounty, props.rangerDistrict, setRangerDistrict);
       setMapClickCallback(() => callback);
       map.on('click', VECTOR_LAYER, callback);
     }
@@ -489,7 +290,18 @@ const PredictionMap = (props) => {
   return (
     <div className="container flex-item-left" id="map-container">
       <div id="map" />
-      <div id="map-overlay-download" onClick={downloadMap}>
+      <div
+        id="map-overlay-download"
+        onClick={() => downloadMap(
+          map,
+          year,
+          isDownloadingMap,
+          setIsDownloadingMap,
+          selectedState,
+          MAP_TYPES.PREDICTION,
+          { titleDetails: { selectedState, period: year }, thresholds, colors },
+        )}
+      >
         <h4>{isDownloadingMap ? 'Downloading...' : 'Download Map'}</h4>
         <div>
           <img id="icon-small"
